week_6_and_7/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
apturl==0.5.2
astroid==3.3.11
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.0
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.23.0
azure-mgmt-core==1.6.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.4.0
azure-mgmt-subscription==3.1.1
babel==2.17.0
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.10.0
bidict==0.23.1
blinker==1.9.0
boto3==1.38.46
botocore==1.38.46
Brlapi==0.8.3
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
cfgv==3.3.1
chardet==4.0.0
click==8.2.1
colorama==0.4.4
command-not-found==0.3
cryptography==45.0.5
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
dill==0.4.0
distlib==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.3.0
Faker==37.1.0
filelock==3.18.0
Flask==3.1.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Compress==1.18
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
httplib2==0.20.2
identify==2.4.10
idna==3.3
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
isodate==0.7.2
isort==6.0.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jmespath==1.0.1
jsonformatter==0.3.4
keyring==25.6.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
ldap3==2.9.1
libgravatar==1.0.4
logilab-common==1.8.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==8.10.0
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
mypy==1.17.0
mypy_extensions==1.1.0
netifaces==0.11.0
nodeenv==0.13.4
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
paramiko==3.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.8.0
pgadmin4==9.5
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==2.17.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.3.7
pymacaroons==0.13.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
pytest==8.4.1
pytest-flask==1.3.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pyxdg==0.27
PyYAML==5.4.1
pyzstd==0.17.0
qrcode==8.2
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
rich==14.0.0
roman==3.3
rsa==4.9.1
s3transfer==0.13.0
SecretStorage==3.3.1
selenium==4.30.0
Send2Trash==1.8.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sshtunnel==0.4.0
systemd-python==234
thonny==3.3.14
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typed-ast==1.4.3
typer==0.16.0
types-aiofiles==0.1
types-annoy==1.17
types-appdirs==1.4
types-atomicwrites==1.4
types-aws-xray-sdk==2.8
types-babel==2.9
types-backports-abc==0.5
types-backports.ssl-match-hostname==3.7
types-beautifulsoup4==4.10
types-bleach==4.1
types-boto==2.49
types-braintree==4.11
types-cachetools==4.2
types-caldav==0.8
types-certifi==2020.4
types-characteristic==14.3
types-chardet==4.0
types-click==7.1
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-contextvars==0.1
types-croniter==1.0
types-cryptography==3.3
types-dataclasses==0.1
types-dateparser==1.0
types-DateTimeRange==0.1
types-decorator==0.1
types-Deprecated==1.2
types-docopt==0.6
types-docutils==0.17
types-editdistance==0.5
types-emoji==1.2
types-entrypoints==0.3
types-enum34==1.1
types-filelock==3.2
types-first==2.0
types-Flask==1.1
types-freezegun==1.1
types-frozendict==0.1
types-futures==3.3
types-html5lib==1.1
types-httplib2==0.19
types-humanfriendly==9.2
types-ipaddress==1.0
types-itsdangerous==1.1
types-JACK-Client==0.1
types-Jinja2==2.11
types-jmespath==0.10
types-jsonschema==3.2
types-Markdown==3.3
types-MarkupSafe==1.1
types-mock==4.0
types-mypy-extensions==0.4
types-mysqlclient==2.0
types-oauthlib==3.1
types-orjson==3.6
types-paramiko==2.7
types-Pillow==8.3
types-polib==1.1
types-prettytable==2.1
types-protobuf==3.17
types-psutil==5.8
types-psycopg2==2.9
types-pyaudio==0.2
types-pycurl==0.1
types-pyfarmhash==0.2
types-Pygments==2.9
types-PyMySQL==1.0
types-pyOpenSSL==20.0
types-pyRFC3339==0.1
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-dateutil==2.8
types-python-gflags==3.1
types-python-nmap==0.6
types-python-slugify==5.0
types-pytz==2021.1
types-pyvmomi==7.0
types-PyYAML==5.4
types-redis==3.5
types-requests==2.25
types-retry==0.9
types-selenium==3.141
types-Send2Trash==1.8
types-setuptools==57.4
types-simplejson==3.17
types-singledispatch==3.7
types-six==1.16
types-slumber==0.7
types-stripe==2.59
types-tabulate==0.8
types-termcolor==1.1
types-toml==0.10
types-toposort==1.6
types-ttkthemes==3.2
types-typed-ast==1.4
types-tzlocal==0.1
types-ujson==0.1
types-vobject==0.9
types-waitress==0.1
types-Werkzeug==1.0
types-xxhash==2.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
virtualenv==20.32.0
wadllib==1.3.6
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.13.3
wsproto==1.2.0
WTForms==3.2.1
xdg==5
xkit==0.0.0
zipp==3.23.0
